+++ server/dbi.js
@@ -27,7 +27,7 @@
  * @since        0.0.0
  * @version      -
  * ************************************************************************** */
-/* eslint-disable one-var, semi-style, no-underscore-dangle */
+/* eslint one-var: 0, semi-style: 0, no-underscore-dangle: 0 */
 
 
 // -- Vendor Modules
@@ -99,5 +99,3 @@ methods = {
 
 // -- Export
 module.exports = DBI;
-
-/* - */