+++ server/libs/sqlite/util.js
@@ -21,7 +21,7 @@
  * @since        0.0.0
  * @version      -
  * ************************************************************************** */
-/* eslint-disable no-underscore-dangle */
+/* eslint no-underscore-dangle: 0 */
 
 
 // -- Vendor Modules
@@ -64,5 +64,3 @@ const Util = {
 
 // -- Export
 module.exports = Util;
-
-/* - */